requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
astunparse==1.6.3
attr==0.3.1
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
black==25.1.0
bleach==5.0.1
blinker==1.9.0
boto==2.49.0
boto3==1.40.22
botocore==1.40.22
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
datamodel-code-generator==0.26.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.11
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
dotenv==0.9.9
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email-validator==2.3.0
expiringdict==1.2.2
Faker==37.6.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.9.0
gast==0.6.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
ijson==3.2.3
importlib_metadata==8.7.0
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsf==0.11.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==3.11.3
kiwisolver==1.4.9
label-studio==1.20.0
label-studio-converter==0.0.59
label-studio-sdk==1.0.18
label-studio-tools==0.0.4
launchdarkly-server-sdk==8.2.1
libclang==18.1.1
lockfile==0.12.2
lxml==6.0.1
lxml_html_clean==0.4.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
openai==1.104.2
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.0.2
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-mock==1.12.1
rich==14.1.0
rpds-py==0.27.1
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.13.1
scipy==1.16.1
semver==3.0.4
sentry-sdk==2.35.2
setuptools==80.9.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
streamlit==1.49.1
streamlit-labelstudio==0.0.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tldextract==5.3.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
ultralytics==8.3.197
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.40.0
wrapt==1.17.3
xmljson==0.2.1
zipp==3.23.0
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
astunparse==1.6.3
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datamodel-code-generator==0.26.1
email-validator==2.3.0
Faker==37.6.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.6
joblib==1.5.2
jsf==0.11.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.2.0
networkx==3.5
nltk==3.9.1
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.4.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==14.1.0
roboflow==1.2.7
rpds-py==0.27.1
rstr==3.2.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.49.1
streamlit-labelstudio==0.0.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.3
xmljson==0.2.1
